pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "larryob"
version = "0.1.0"
description = "NBA analytics database: SQLite ingestion (OLTP) + DuckDB analytics (OLAP)"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "duckdb>=1.4.4",
    "nba-api>=1.11.4",
    "pandas>=3.0.1",
    "requests>=2.32.5",
    "lxml>=6.0.2",
    "pydantic>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
ingest = "src.pipeline.cli.main:main"

[project.optional-dependencies]
analytics = [
    "pyarrow>=23.0.1",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
    "ty>=0.0.18",
    "pytest-cov>=7.0.0",
    "types-requests",
    "mutmut>=3.4.0",
    "pyyaml>=6.0.3",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]  # line-too-long handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests
"tests/test_etl_helpers.py" = ["S101"]
"tests/test_etl_dimensions.py" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]

